Unit test admin voucher works

+++ Project/DataModels/VoucherModel.cs
@@ -3,13 +3,13 @@ public class VoucherModel
     public Int64 Id { get; set; }
     public string Code { get; set; }
     public string Description { get; set; }
-    public int Amount { get; set; }
+    public decimal Amount { get; set; }
     public string Type { get; set; }
     public int? UserId { get; set; }
 
     public VoucherModel() { }
 
-    public VoucherModel(Int64 id, string code, string description, int amount, string type, int? userId = null)
+    public VoucherModel(Int64 id, string code, string description, decimal amount, string type, int? userId = null)
     {
         Id = id;
         Code = code;
